refactor: pr #25 리뷰 반영
