+++ pages/auth/cb/facebook/register.tsx
@@ -1,5 +1,5 @@
 import { useStringParam } from '@/components/auth/hooks';
-import { useFacebookAuth } from '@/components/auth/idp/facebook';
+import useFacebookAuth from '@/components/auth/idp/useFacebookAuth';
 import { useRouter } from 'next/router';
 import { FC, useState } from 'react';
