+++ pages/auth/register.tsx
@@ -1,4 +1,4 @@
-import { useFacebookAuth } from '@/components/auth/idp/facebook';
+import useFacebookAuth from '@/components/auth/idp/useFacebookAuth';
 import { useStringParam } from '@/components/auth/hooks';
 import axios from 'axios';
 import { FC, useState } from 'react';
